orders/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiokafka==0.12.0
aiomysql==0.2.0
aiosignal==1.3.2
aiosmtplib==4.0.1
aiosqlite==0.21.0
alembic==1.15.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.1.0
Authlib==1.5.1
bcrypt==4.3.0
billiard==4.2.1
bson==0.5.10
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==44.0.2
dnspython==2.7.0
elastic-transport==8.17.1
elasticsearch==8.13.2
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
firebase-admin==6.9.0
frozenlist==1.5.0
pyzbar==0.1.8
pillow>=9.0.0
qrcode==7.3.1
colorama==0.4.6
python-barcode==0.14.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
ksuid==1.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.1
msgpack==1.1.1
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
prompt_toolkit==3.0.51
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psycopg==3.2.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.2
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
setuptools==76.1.0
shellingham==1.5.4
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
yarl==1.18.3
